+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Unreleased
 - lib.platforms.director.DirectorV4Parser: Add code dumper
 - fields.Bits: Fix 32 and 64 bit support
 - lib.games.sam: Fix typo
+- lib.platforms.director: Fix implementation of ScriptFloat
 
 0.7.3 - 2019-12-07
 ==================